modularizado
