+++ api/src/Controllers/Products/createProduct.js
@@ -1,4 +1,4 @@
-const { Products } = require("../db")
+const { Products } = require("../../db")
 const createProduct = async ({name, category, brand, images, description, price, avalaible, stock }) => {
     const addProduct = await Products.create({name, category, brand, images, description, price, avalaible, stock });
     return addProduct;
