+++ api/src/Controllers/Products/findById.js
@@ -1,4 +1,4 @@
-const { Products } = require("../db")
+const { Products } = require("../../db")
 
 const findProduct = async (idProduct) => {
       const productById = await Products.findByPk(idProduct);
