+++ api/src/routes/productsRouter.js
@@ -1,7 +1,7 @@
 const express = require("express");
-const findAllProducts = require("../Controllers/findAllProducts")
-const findProductByName = require("../Controllers/findProductByName")
-const findById = require("../Controllers/findById")
+const findAllProducts = require("../Controllers/Products/findAllProducts")
+const findProductByName = require("../Controllers/Products/findProductByName")
+const findById = require("../Controllers/Products/findById")
 const {Products} = require('../db')
 
 const productsRouter = express.Router()
